requirements.txt
# Core Framework
streamlit>=1.46.1
pandas>=2.0.0
numpy>=1.24.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
pydantic>=2.0.0

# HTTP & API
requests>=2.31.0
httpx>=0.25.0
feedparser>=6.0.0

# Data Visualization
plotly>=5.17.0

# Authentication (Basic)
streamlit-authenticator>=0.2.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Future dependencies (commented out for now)
# fastapi>=0.104.0
# uvicorn>=0.24.0
# sqlalchemy>=2.0.0
# redis>=5.0.0
# langchain>=0.1.0
# openai>=1.3.0
# anthropic>=0.7.0
